// apps/web/src/app/core/models/ApplicationApiKeysModel.ts
export interface IApplicationApiKeys {
  applicationApiKeyId: string;
  applicationId: string;
  organizationId: string;
  environment: Environment;
  keyHash: string;
  keyPrefix: string;
  keyType: ApplicationApiKeyType;
  permissions?: string[];
  status: ApplicationApiKeyStatus;
  nextKeyHash?: string;
  activatesAt?: Date;
  expiresAt?: Date;
  revokedAt?: Date;
  lastUsedAt?: Date;
  createdAt: Date;
  updatedAt: Date;
  ttl?: number;
}